created user_feeback doc with Tonality and SynthFeedback classes
