+++ run.py
@@ -39,7 +39,8 @@ if args.app == 'backend':
 elif args.app == 'song':
     from pythonosc import udp_client
     from song import song_machine
-    from song.song_server import SongServer, BeatAdvanceManager, Tonality
+    from song.song_server import SongServer, BeatAdvanceManager
+    from song.user_feedback import Tonality, SynthFeedback
 
     oscul_client = udp_client.SimpleUDPClient(settings.ip, settings.OSCULATOR_PORT)
     audience_client = udp_client.SimpleUDPClient(settings.ip, settings.AUDIENCE_PORT)
@@ -47,9 +48,11 @@ elif args.app == 'song':
 
     machine_instance = song_machine.create_instance(settings.song_path)
     tonality = Tonality(machine_instance.parser.categories)
+    synth_feedback = SynthFeedback(machine_instance.parser.categories)
     beat_manager = BeatAdvanceManager(machine_instance.current_part)
 
-    song_server = SongServer(oscul_client, audience_client, performer_client, machine_instance, beat_manager, tonality)
+    song_server = SongServer(oscul_client, audience_client, performer_client, machine_instance, beat_manager,
+                             tonality, synth_feedback)
     song_parts = list(machine_instance.parser.song_parts.keys())
     [audience_client.send_message('/parts', part) for part in song_parts]
     audience_client.send_message('/parts', 'all_sent')
